fix(settings): address search review
